Config file for game settings (#17)
* Initial config file implementation

* Log unexpected key and use systems

Uses only ECS systems now to tidy it up. Also will log a warning if
there is an unexpected key in the config file.

* Add clippy lints

Add `semicolon_if_nothing_returned` and `uninlined_format_args` for more
consistent code style

* Update README.md

* Add `shell.nix`

* Add instructions and feature for wayland

* Remove config plugin

We don't need to pretend that we can change the config while the game is
running (yet).